introduce simpler ClientResultError instead of MonadError
